requirements.txt
fastapi~=0.115.0
starlette~=0.38.6
pydantic~=2.9.2

tortoise-orm~=0.21.6
asyncpg~=0.29.0

uvicorn[standard]~=0.30.6

passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
python-multipart~=0.0.9
python-dotenv~=1.0.1